SDK: switch to musl-static + RISC-V SBI vendor ABI
This is a foundational rewrite of the openfpgaOS SDK and the kernel↔app
boundary. The libc jump table is gone; apps now statically link unmodified
upstream musl 1.2.5 and emit Linux RISC-V syscalls, while openfpgaOS-specific
subsystems live in an SBI-style vendor extension space.

Why: the libc jump table caused recurring stability issues, was a fork
point we owned, and prevented LTO from stripping unused musl symbols.
The new model is what every shipping musl distribution does (Alpine, Void),
gives us zero ongoing libc maintenance, and matches the standard SBI
calling convention so anyone reading the code with RISC-V firmware mental
models recognizes it instantly.

Highlights
==========

* SBI-style vendor ABI for openfpgaOS subsystems
  - a7 = EID, a6 = FID, a0..a5 = args, returns sbiret { error, value }
  - Vendor base 0xC0DE0000 (high bit set; cannot collide with Linux
    syscall numbers or future standard SBI extensions)
  - Per-subsystem FID enums in api/of_syscall_numbers.h, append-only
  - 15 EIDs + ~80 FIDs covering video/audio/input/mixer/timer/net/...
  - of_error.h SBI-aligned (-1..-13 match SBI spec exactly)
  - Trap handler in start.S loads both a6/a7, calls C dispatcher,
    decides whether to write back a1 based on EID range so the Linux
    path preserves the user's a1 (musl declares only a0 as clobbered)
  - Two-level switch dispatcher in kernel/syscall.c, replaces the old
    if-chain that had the mixer overflowing its 16-slot zone
  - docs/syscall-abi.md: full specification + extension recipes

* Apps statically link upstream musl 1.2.5
  - Deleted libc_include/, of_libc.h, of_posix.c, of_malloc.c
  - Deleted kernel/libc_table.c (and its main.c init call)
  - Moved dlmalloc.c into kernel/ (kernel-only now)
  - of_caps.h: dropped libc_table field
  - SDK app build (api/sdk.mk + api/app.ld) uses musl crt1.o + crti.o +
    crtn.o, links libc.a + libm.a + libgcc, no toolchain libsupc++
    (newlib-targeted, can't be used against musl)
  - api/of_cxxabi.cpp: minimal C++ ABI shim (operator new/delete,
    __cxa_pure_virtual, __cxa_atexit, static guards) for cxxdemo and
    other C++ apps
  - api/Makefile.fpga: standalone test template
  - musl/build_musl.sh: hardened to verify crt1.o/crti.o/crtn.o get
    installed; previous build was producing libc.a only
  - Result: os.bin shrank from ~122 KB to ~64 KB (LTO can finally
    strip unused musl functions)

* Loader auxv now compliant
  - kernel/loader.c::elf_exec() pushes AT_PAGESZ=4096, AT_HWCAP=0,
    AT_RANDOM=NULL, AT_NULL terminator
  - Without AT_PAGESZ, libc.page_size stayed 0 and mallocng silently
    divided by zero on rv32, hanging any app that called malloc
    (testdemo's first malloc-using test). This was the "demos that
    don't malloc work, demos that malloc hang" symptom.

* SDK header cleanup
  - Removed deprecated of_link.h, of_bram.h, of_video_sync(),
    of_mixer_set_vol_rate(), OF_HW_LINK, OF_PLATFORM_DE10NANO
  - Renamed kernel-side of_mixer_set_vol_rate -> of_mixer_set_volume_ramp
  - Added boolean enum constants (OF_ENABLE/OF_DISABLE/OF_LOOP/etc)
  - Added of_video_wait_flip (proper name for what was of_video_sync)
  - Fixed stale comments + counts in of_libc/of_caps/of_midi
  - of.h: switched to canonical of_net.h / of_fastram.h includes,
    uses OF_SYS_EXIT instead of magic 93
  - of_terminal.h, of_timer.h: declared previously-hidden symbols
  - of_input.h: documented single-TU constraint loud and clear

* Build hardening
  - sdk.mk + Makefile.fpga + Makefile.pc + os/Makefile all add:
      -falign-functions=64 -falign-loops=64
      -fno-tree-loop-distribute-patterns
  - tools/sweep.sh: tolerates timing-not-met seeds, picks best Fmax
    regardless of slack closure; sweep no longer aborts on Quartus's
    non-zero exit code, verifies .sof presence as the real success
    criterion; prints a yellow warning when winning seed is below
    100 MHz so we still know what we're shipping
  - ap_core.qsf default seed: 21

* Sample memory cleanup (cosmetic)
  - hal/mixer.c: extracted bump allocator into its own section

* Documentation
  - docs/syscall-abi.md: full SBI ABI spec
  - docs/app-development.md: rewritten Standard C Library section,
    Build Requirements pointing at Makefile.fpga, full Linux syscall
    table, removed phantom of_print_char/clear/at and of_save_*
    functions that never existed
  - docs/developer-guide.md: stale "libc jump table" references
    replaced; phantom function entries removed
  - docs/mister-support-proposal.md: new

+++ src/firmware/os/targets/pocket/save.c
@@ -50,7 +50,7 @@ void of_save_init(void) {
 }
 
 int of_save_read(int slot, void *buf, uint32_t offset, uint32_t len) {
-    if (slot < 0 || slot >= SAVE_MAX_SLOTS)
+    if (slot < 0 || slot >= (int)SAVE_MAX_SLOTS)
         return -1;
     if (offset + len > SAVE_SLOT_SIZE)
         return -1;
@@ -64,7 +64,7 @@ int of_save_read(int slot, void *buf, uint32_t offset, uint32_t len) {
 }
 
 int of_save_write(int slot, const void *buf, uint32_t offset, uint32_t len) {
-    if (slot < 0 || slot >= SAVE_MAX_SLOTS)
+    if (slot < 0 || slot >= (int)SAVE_MAX_SLOTS)
         return -1;
     if (offset + len > SAVE_SLOT_SIZE)
         return -1;
@@ -78,7 +78,7 @@ int of_save_write(int slot, const void *buf, uint32_t offset, uint32_t len) {
 }
 
 int of_save_flush_size(int slot, uint32_t size) {
-    if (slot < 0 || slot >= SAVE_MAX_SLOTS)
+    if (slot < 0 || slot >= (int)SAVE_MAX_SLOTS)
         return -1;
     if (size > SAVE_SLOT_SIZE)
         size = SAVE_SLOT_SIZE;
@@ -97,7 +97,7 @@ int of_save_flush(int slot) {
 }
 
 void of_save_update_crc(int slot) {
-    if (slot < 0 || slot >= SAVE_MAX_SLOTS)
+    if (slot < 0 || slot >= (int)SAVE_MAX_SLOTS)
         return;
 
     volatile uint8_t *base = slot_base(slot);
@@ -109,7 +109,7 @@ void of_save_update_crc(int slot) {
 }
 
 int of_save_check(int slot) {
-    if (slot < 0 || slot >= SAVE_MAX_SLOTS)
+    if (slot < 0 || slot >= (int)SAVE_MAX_SLOTS)
         return -1;
 
     volatile save_meta_t *meta = slot_meta(slot);
@@ -131,7 +131,7 @@ uint32_t of_save_get_size(int slot) {
 }
 
 void of_save_erase(int slot) {
-    if (slot < 0 || slot >= SAVE_MAX_SLOTS)
+    if (slot < 0 || slot >= (int)SAVE_MAX_SLOTS)
         return;
 
     volatile uint32_t *dst = (volatile uint32_t *)slot_base(slot);
